+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 All notable changes to this project will be documented in this file.
 We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
+## Unreleased
+- Load validation scenarios from external file instead of Python module.
+
 ## 0.9.95 - 2017-03-17
 - Enforce non-zero price in ratechecker API.
 - Add new ratechecker API endpoint to check data load timestamp.